requirements.txt
# Django stuff
Django==1.2.1
django-annoying==0.7.4
-e git://github.com/fwenzel/django-mozilla-product-details#egg=django-mozilla-product-details

# DB stuff
MySQL-python==1.2.3c1
-e git://github.com/jbalogh/schematic.git#egg=schematic

# Term extraction
topia.termextract==1.1.0

# Search
django-haystack==1.1.0-alpha
Whoosh==0.3.18
